clever-data-jdbc

+++ clever-data-jdbc/src/main/java/org/clever/data/jdbc/mybatis/AbstractMyBatisMapperSql.java
@@ -197,7 +197,7 @@ public abstract class AbstractMyBatisMapperSql implements MyBatisMapperSql {
     }
 
     @Override
-    public void startWatch(int period) {
+    public void startWatch(long period) {
         if (watch) {
             return;
         }
